refactor: 야구게임, 힌트 클래스 분리
